fix: interface pollution for service and index
